% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_sd.R
\name{calc_mean_sd}
\alias{calc_mean_sd}
\title{Calculate Mean and Standard Deviation}
\usage{
calc_mean_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing the mean and standard deviation.
}
\description{
Computes the mean and standard deviation of a numeric vector.
}
\examples{
calc_mean_sd(c(1, 2, 3, 4, 5))
}
